更新"特伍拉汽車工廠"的程式碼

+++ 特伍拉汽車工廠/BasicCar.cs
@@ -20,12 +20,12 @@ namespace 特伍拉汽車工廠
 
         // Method
         
-        public override double GetCost()
+        public override int GetCost()
         {
             return eng.GetCost() + ac.GetCost() + 5000; // 引擎成本 ＋ 空調成本 + 5000
         }
 
-        public override double GetPrice()
+        public override int GetPrice()
         {
             int price = (int)Math.Round(GetCost() * 1.2);
             return price;
